Split some tests out of existing_data/mod.rs
